// local-hobbies-frontend/src/types/availability.ts
export enum AvailabilityFrequency {
    Weekly = 'Weekly',
    EveryTwoWeeks = 'Every two weeks',
    Monthly = 'Monthly',
}

export enum DayOfWeek {
    Monday = 'Monday',
    Tuesday = 'Tuesday',
    Wednesday = 'Wednesday',
    Thursday = 'Thursday',
    Friday = 'Friday',
    Saturday = 'Saturday',
    Sunday = 'Sunday',
}
// ...
export interface RecurringAvailabilityCreationRequest {
    location: GeoJsonPoint;
    /** ISO 8601 date (YYYY-MM-DD) */
    ruleStart: string;
    /** ISO 8601 date. Nullable if rule continues indefinitely */
    ruleEnd?: string | null;
    frequency: AvailabilityFrequency;
    startDayOfWeek?: DayOfWeek | null;
    startDayOfMonth?: number | null;
    /** ISO 8601 time (HH:mm) */
    startTime: string;
    /** ISO 8601 duration */
    duration: string;
}

export interface RecurringAvailabilityUpdateRequest {
    location?: GeoJsonPoint;
    /** ISO 8601 date */
    ruleStart?: string;
    /** ISO 8601 date */
    ruleEnd?: string | null;
    frequency?: AvailabilityFrequency;
    startDayOfWeek?: DayOfWeek | null;
    startDayOfMonth?: number | null;
    /** ISO 8601 time */
    startTime?: string;
    /** ISO 8601 duration */
    duration?: string;
}

export interface RecurringAvailabilityResponse {
    id: string;
    location: GeoJsonPoint;
    /** ISO 8601 date */
    ruleStart: string;
    /** ISO 8601 date */
    ruleEnd: string | null;
    frequency: AvailabilityFrequency;
    startDayOfWeek: DayOfWeek | null;
    startDayOfMonth: number | null;
    /** ISO 8601 time */
    startTime: string;
    /** ISO 8601 duration */
    duration: string;
}